add protocol to allow content sources other than secret
A backward compatible implementation to allow specify a protocol where
only a secret name was being used. If, instead a secret name a
`proto://content` is used, such content will be searched based on the
protocol. If a protocol isn't provided, `secret` is used as default.

Actually only `file` protocol is implemented which expects an absolute
path, but others like `cm/configmap`, `http/s` can also be implemented
in the future.

Internal names wasn't changed and so the public interface - currently
all namings reference "secret". Docs updated where the protocol can be
used.