improve crt validation with ssl_c_verify
HAProxy Ingress validates client certificates using ACL in order to
give a http response. The samples used were `ssl_c_ca_err` that returns
>0 if there is a problem in the ca bundle (depth>0), and `ssl_c_err`
that returns >0 if there is a problem in the certificate itself.
However, there were observed outdated client certificates that doesn't
trigger a failure in `ssl_c_err` but it does in `ssl_c_verify` which
also check against the CA bundle and motivated this update.