protect show proc parsing against crashes
HAProxy's master unix socket is used to read process state, usually in
the event of a reload and launching a new worker. The output has lines
with fixed length and fields with fixed starting positions, but
eventually an output can be truncated. This update prevents a crash,
checking the length of the buffer before cutting a field from it.