read the whole input when the response fills the buffer
HAProxy responses has just a few bytes, usually saying that a backend
server was successfully updated. A 1k buffer is just enough to handle
such responses. Since the external process implementation, one of the
responses has a list of all running procs, which will easily fill that
buffer depending on the number of old workers alive. This updates the
func to continue reading the buffer up to the last byte.