remove unix socket before start acme server
An unix socket is used to answer acme challenges - it is create whenever
--acme-server command-line is enabled. However it isn't removed if
haproxy ingress is abruptly stopped, eg when a crash occurs. Since v0.12
an external volume is created if haproxy is running as a sidecar,
leaving the unix socket file alive between restarts, causing a
crashloop in the controller if the acme server is enabled. This update
ensures that the unix socket file isn't present when the server is
started