Fix initial weight configuration
Weights can be used to distribute distinct load to distinct servers in
the same backend. An initial value greater the default 1 (one) can be
configured so agents that changes server weight can properly work and
adjust the weight in lower granularity.

The initial weight configuration isn't properly working when declared in
the ingress resource because its usage occurs before the value is
actually read from the ingress annotation. It works however when
declared in the service - service has priority and is read in the same
block the backend is created and configured with initial weight and
other defaults. Moving down the reading just before starting the
endpoint configuration fixes this.