add external authentication
This update adds the ability to configure an URL used to authenticate a
request. Every single request passes through auth-request.lua ->
haproxy-lua-http.lua -> frontend listening a local tcp port -> target.
The local port uses an unassigned range by default and can be
configured. An optional signin redirect can be configured in the case
of an authentication failure, useful with oauth2.