fix backend matches if hostname uses wildcard
Some hostname+path matches happen in the backend to identify if a
configuration should or should not be applied. Examples are whitelist,
cors, hsts and rewrite path. Frontend matches are correctly using a
regex if a wildcard hostname is used. Backends also work but only if all
paths share the same configuration, in this case a path ACL is not used.
However backends that have distinct path configurations were not
properly configuring regex based ACLs - they were trying to match
wildcard hostnames with -m begin.

This update adds to the backend the same hostname check used in the
frontend, which converts begin match to a regex match if a wildcard
hostname is used.