add server redirect options
A server redirect allows to configure source domains that, if provided
in a request, should redirect to the domain configured in the ingress'
spec. Valid hostnames will always have precedence, which means that in
the case of a match, the redirect will not happen.

As a host scoped config, two new maps were created (redirect source and
status code) which will only be used if the functionality is also used.