move oauth configuration to auth-url and signin
Reuse the internal types that supports `auth-url` and `auth-signin` on
behalf of the oauth configuration keys. This update is backward
compatible with the oauth2 internal types and haproxy templates, except
the compatibility changes already documented when `auth-reqeust.lua` was
updated.

OAuth2 and auth external cannot coexist, if both are configured in the
same path, oauth2 will be ignored.