add https and svc proto to auth-url config key
Adds https and svc protocols. HTTPs was built on top of HTTP, only
changing backend's secure (true), adding sni (if a hostname was
provided) and default port if not provided (443). Server crt is not
validated, leading to a less secure connection.

Svc adds the ability to reference a service which can be annotated
to provide a properly secured connection with CA and CRL.