remove unused helper front/backend for auth-url
A proper tracking would be more appropriate here, but the current
tracking implementation doesn't allow us to go further. This
implementation 1) removes the front/back helper that links to a backend
used with svc protocol which is not used anymore, avoiding "backend not
found" erros from haproxy; and 2) Removes unused front/back helpers if
acquiring a new helper lead to "auth proxy list is full".