fix default host as a ssl-passthrough
SSL passthrough domains are added in the same hostname maps used to
route http requests. They share most of their configurations with a few
exceptions. One of these exceptions weren't properly implemented -
default host was always being configured in the http proxy. This update
checks if the default host is an ssl passthrough, properly configuring
it in the tcp proxy instead.