add ann parsing of default tcp host
TCP services support SNI inspection and has its default host, pretty
much like the HTTP one - it configures the backend that should be used
if there is no matching host. This update adds annotation parsing from
the default host, and does not have impact in the codebase because,
currently, there is no hostname scoped annotation for TCP Services. A
fix for a future problem.