add non root path support on ssl-passthrough
A ssl-passthrough host cannot inspect uri and http headers because the
request isn't offloaded in haproxy - a TCP tunnel is created instead.
This however isn't true for the plain http side of the configuration,
which up to now can only be configured with a 302 redirect (default
behavior) or a plain http port in the same backend.

This update adds the ability to configure non root paths. Such paths
does not change how the https side behaves, but do so for the http one.
`ssl-passthrough-http-port` config key continue to be supported in the
same way, but now act as a fallback - how the request will be handled
in the http side if a path isn't found.