allow to configure a list of annotations prefix
This update converts the annotations prefix to a slice, allowing to
configure more than one supported prefix. Using a list of prefix makes
it possible to conveniently change the default value without breaking
backward compatibility. In the case of a conflict, the prefix that
appears first in the command-line option will have precedence.