fix reading of needFullSync status
`needFullSync` is a k8s cache status flagged when a changed resource
cannot be properly identified. This status instructs HAProxy Ingress to
parses all the ingress resources and build a new haproxy model from
scratch, instead of try to perform a partial parsing.

The status was being read just after a cleanup, so a full sync wouldn't
run if required. This update moves the current status to the changed
objects struct just before being cleaned up.