add initial implementation of gateway api converter
Initial Gateway API converter implementation with a single
functionality, plenty of todos along the way, no error handling, no
partial parsing and no unit tests.

The main purpose of this change is to run an e2e test, validate listers
and notifiers, and also make a small refactor on ingress and configmap
based tcp services, allowing to insert gateway in the converter pipeline
and reuse the actual codebase.