fix per path filter of default host rules
This update fixes two issues in rules with default host (empty hostname)

The first one is in the backend, the per path matching was trying to
match the hostname with an internal representation. In this case only
the path should be matched after the non empty hosts has been failed.

The other one is in the haproxy's frontend, the path was being matched
with path_beg despite of the path type configuration. Now default host
match uses a map just like non empty hosts.