fix path-type conflict warning
Scopes are used to identify and log configuration conflict, and also to
group configurations related to the same annotation mapping. Path type
was wrongly configured as host scope and, although it was working as
expected, a conflict warning was being logged if using two distinct
path types in the same hostname. Change to path scope removes the
warning without changing the behavior.