give the maximum priority to the exact path type
Exact path type has a special meaning in the map building: it's the
faster match and it doesn't overlap with any other match type and even
with other exact paths. This update gives the highest priority to the
haproxy maps of this type, and also fixes a use case where the same
path is added as exact and eg as prefix: exact should be checked first
or it will never match.