accept header names in auth and oauth-headers
`auth-request.lua` sanitizes header names from the authentication
service in a predictable way. So instead of document this to the user,
we're now running the same sanitization in code, simplifying the user
interface with the configuration key. Now `auth-headers` and
`oauth-headers` can configure just the header name (the same name is
copied from the auth service), another header name as the source, as
well as a HAProxy variable name. The later gives some flexibility and
preserve backward compatibility. Vars are identified from the prefix
used in the name, and if identified, their names are used verbatim
instead of performing the sanitization.