fix reference of gateway listeners
All gateway listeners are iterated and filtered, looking for distinct
listener types and filling an array of httproutes. Future implementation
will have also tcproutes and udproutes. The resulting slice receives a
pointer to the listener, but was incorrectly referencing the variable
used to iterate the source list. Now the list is iterated using its
index, so the reference is correctly assigned to the list item.