add ssl-always-add-https config key
Since at least v0.8 controller adds domains in both HTTP and HTTPS maps
despite the use of the `tls` attribute. This behavior goes against the
Ingress spec but cannot be simply fixed due to backward compatibility.
This config key adds the ability to configure HTTPS without the need of
the `tls` attribute and can be globally changed. Its default value is
`true` to preserve the behavior of current and older versions.

Should be merged to v0.12.