use the port name on DNS resolver template
A haproxy backend naming uses the service namespace and name, and also
the port of the service. Kubernetes has a number of ways to reference a
port, which can be a number or a name, and both can be of the service
itself or the target (a pod). Current implementation just use the
target reference, and being a number make it in sync with the port
number in the target server. Anyway this is just a naming convention,
except when DNS resolvers are in place.

The DNS resolver feature expects the name of the service port because
this is the name the DNS knows. When port name and target port differs,
the server-template is wrongly built with the target name. This commit
adds a new DNSPort attribute that the DNS template should use to build
the correct configuration.