ensure that configured global ConfigMap exists
Launch is actually copying the global configmap name from the
command-line straight to the lister without checking its existence. The
lister will watch configmap events and will only wake up if the name
matches with the one provided in the command-line, so currently a
mistyped name is silently ignored.

From now the configmap name is checked in the bootstrap process and will
crash if not found or if the controller doesn't have permission to read
it.

This should be merged as far as v0.10, but v0.12 and older should only
log error instead, due to preserve backward compatibility with the
current behavior.