reload haproxy if a backend server cannot be found
The controller maintain and evolve an in memory model that reflects the
haproxy state. This model is used to rebuild the configuration files and
also to name haproxy backends and servers when running a dynamic update.
If for any reason the model and the haproxy instance loose their
synchronization, backends or servers might not be found and a dynamic
update would fail. This update parses the haproxy response, if it
returns a `No such server` or any other non recognized message, haproxy
will be fully reloaded in order to resynchronize with the model.