add reload interval command-line options
Creates a reload queue that manages how frequently haproxy can be
reloaded. A mutex ensures that the ingress queue and converter updates
doesn't conflict with the model reading from the haproxy reload queue.
The queue is only used if the reload interval option is greater than
zero, so it's a safe addition on its default configuration.