change nbthread to use all CPUs by default
Removes the default value of 2 threads and let HAProxy configures
threads automatically if the configuration is not provided. It's
preferable more than less threads on a poorly configured system (or
benchmark).

Compatibility notes: this might change the number of configured threads
on deployments that doesn't configure an explicit value. The number
should be updated to the number of CPUs available to the HAProxy
process, or to one thread if the platform doesn't support CPU affinity.