add keep alive false option
Add the ability to disable keep alive (prompt mode, interactive) when
creating HAProxySocket. This will be the default to the current/living
instance - if for any reason HAProxy is restart outside the controller,
the established connection with the old instance would be used to send
dynamic updates. This can be improved in the future. The same connection
is still being used if more than one command is sent at the same time,
which happens when enabling and disabling backend servers.