docs: add section for AuditLog sidecar for ModSecurity daemonset
In our customization of the ModSecurity agent, we have certain rules
triggered to write to the auditlog and ignore being printed to the
modsecurity-spoa container STDOUT. From my research, there doesn't
appear to be a way to configure ModSecurity to also print the audit log
file to STDOUT, but that might be better anyways because it would make
it harder to read a file being edited with two different log sources.

We need to view what's going into the audit log file through container
STDOUT. When we created the daemonset, we added a sidecar container to
print that file to STDOUT for further ingestion.

Even though it is simple, it is extremely useful if someone was to
seperate the destinations of logs for different rules and actions.