fix dynamic update of frontend crt
Since v0.13 certificates can be dynamically updated, however a reload
was always scheduled due to an indirect reference from the backend to
the frontend's certificate hash. This hash is used to the backend know
if the frontend has a TLS certificate. When the certificate is changed,
the hash is changed as well, leading to a configuration change that
triggers a reload. The reference was moved to the filename, which can
also be used to check TLS enabled and cannot be changed.