docs: changing NodeSelector to ClusterIP service for ModSecurity (#826)
Adding more to the ModSecurity docs to show the configuration of
ModSecurity as a ClusterIP service wihout a NodeSelector. This way you
can have the daemonset schedule new pods whenever the cluster scales
its nodes, without having to label the new nodes or adding additional
endpoints for the new pods.