fix endpoint update of configmap based tcp services
Legacy ConfigMap based TCP services has its own endpoint list that
doesn't track partial parsing. The entry point was changed when Gateway
API was added, it was was incorrectly configured to only run when a full
sync is needed, or when the ConfigMap was changed, leading to new state
not being reflected in the configuration. This update ensures that
ConfigMap based TCP services is always parsed when something changes in
the cluster.