fix validation if ca is used with crt and key
An undocumented but inherited behavior from the so called ingress
controller common code base allows to use a ca.crt key along with
crt and key. However a crt validation was failing due to the missing
intermediate crts in the validation. There is no behavioral change
when using crt and key without a ca.