always deny requests if oauth is misconfigured
An attempt to enable oauth should end up with a proper configuration,
or the request should always fail in order to protect the backend
service against misconfigurations. This change ensures that an attempt
to configure oauth will lead to a proper configuration, or all requests
will be denied.

This already works this way on v0.13, should be merged between v0.12
and v0.10.