docs: add modsec resource limits to controls V2 memory consumption (#841)
From our deployment/daemonset of ModSecurity, we have been noticing the
ModSecurity pods hogging memory after our performance tests finish.
There is no garbage collection done and the memory consumed only
increases over time.

We believe this is due to documented memory leak issues with V2 of
ModSecurity. Our kubernetes native solution to this is to provide
memory allocations to the deployment, to force restarts of the pods
instead of hogging memory on nodes. We've ran performance tests against
this change to the deployment and there is no performance regression.