refactor tracker to an abstract implementation
Tracker is used to link kubernetes resources and haproxy entities each
other, allowing to list, remove and rebuild all objects related with a
configuration change. This is the base of the partial parsing
implementation.

The codebase used in v0.11 (starting implementation) up to v0.13 is
a lot bigger than necessary, coupled with k8s and haproxy types , and
also tricky to extend and test. This update changes the code base to an
abstract implementation that can be expanded without the need to write
more code, but instead just declaring a new type to every new k8s or
haproxy type that need to be tracked.

Both old and new code base are compatible, they only change in the
interface declaration, and the new one doesn't have a special behavior
when the target doesn't exist - being missing or not, a resource
creation or update will trigger the list+remove+recreate in the same
way.