separate service and endpoints tracking
Kubernetes Services and Endpoints share the same resource name and
behavior from the controller point of view - no matter who changed,
controller will behaves in the same way. Since the first tracking
version Services and Endpoints are managed in the same scope.

This update creates a new scope to the Endpoints resource. This fixes
event logging and also allows to track Services and Endpoints
individually.