fix global config-backend
The mitigation of CVE-2021-25740 added a command-line option to allow a
sysadmin to partially or completely remove backend configuration
snippet. This option added a new code where there was only a
straightforward assignment. This code fails on how to check if a
configuration has a source and, instead of adjust the logging message,
the whole assignment was being skipped. However a global configuration
by definition doesn't have a source, leading to backend configuration
snippet being skipped if not configured via ingress or service
annotation.

This fix should be merged up to v0.10.