fix error message on secret/cm update failure
Acme functionality needs permission to create or update secrets when a
new certificate is issued, and also to save the client's private key.
This permission isn't added in the default configuration. The func that
creates or updates the secret should return an error if anything went
wrong, but the variable that has the error instance was being redeclared
inside the if block, and was always returning nil even if an error (eg
missing permission) occur.

The same is happening with ConfigMap, which is used to save the
challenge token.