add disableKeywords only if defined
`--disable-config-keywords` command-line option allows to define a
comma separated list of blacklisted keywords. If not declared, an
empty string is used as an input to `Split()`. This empty string is
however converted in an array of one element with an empty string, and
such empty keyword matches with empty lines, denying configurations with
empty lines. This update ignores empty entries when parsing the custom
snippet, and also `TrimSpace()` all the options in order to properly
match with the custom snippet tokenizer.