add match method on all var() sample fetch method
Historically var() sample fetch used to return string so a match method
was optional. Currently var() returns undefined, and since 2.5 the
match method is mandatory. This helps to not make bogus configuration,
when e.g. an int type is expected but it's in fact matched as string.
See https://www.mail-archive.com/haproxy@formilux.org/msg41341.html

This update should be merged as far as v0.12, which is the first
controller version that supports external haproxy.