add master-worker mode on embedded haproxy
add the ability to run the embedded haproxy in master-worker mode.
Master-worker runs in the foreground, allowing to control the lifecycle
of the process in a dedicated goroutine, within an event loop that will
relaunch the process in the case of an unexpected crash.

Default behavior remains the same, so daemon mode will continue to be
used if no configuration option is added, this should probably be
changed in a future version.