change auth-tls-scrict from false to true
`auth-tls-scrict` is used to define whether a misconfigured TLS based
authentication should always fail, i.e. defines if misconfigured should
ignore the configuration or create a fake configuration to protect the
backend.

Historically HAProxy Ingress alerts and discards wrong configuration
but this could lead to a security issue on configurations that protect
applications. Strict option was added in v0.8 as false to preserve
backward compatibility, and finally changed its default to true to
improve security.