check by default if gateway api crds are installed
Change default value of `--watch-gateway` command-line option to true,
so try to listen to gateway API CRDs by default. This new implementation
will skip registration of listers and informers if the crds are not
installed in the server, contrary to v0.13 which would always try to register them.

There is a backward incompatibility, a v0.13 controller that doesn't
declare the command-line option would ignore gateway api resources,
while v0.14 would find and configure them by default.