suffix gateway API v1alpha1 related names with A1
Renaming all Gateway API related identifiers to a distinct name that
does not collide with the new version. This will allow us to process
both v1alpha1 and v1alpha2 in the same controller, helping users to
migrate from these two incompatible CRDs.

In addition only listers and informers being used were maintained:
gatewayclass, gateway and httproute.

This is only an internal refactor, there is no change in the behavior.