add haproxy 2.5 support on external haproxy
Creates a new `show proc` parser compatible with haproxy 2.2 to 2.5.
Added in a safe way to allow merging as far as controller v0.12 -
version 2.4 and below will continue to use the old parser, minimizing
the risk of adding an unknown bug from the new one.

Should be merged as far as v0.12, which is the first version that
implements external haproxy and parses `show proc` output.