fix match of prefix pathtype if using default host
HAProxy's subdir match type is used to implement the `Prefix` pathType.
This match type however doesn't implement exactly what is expected from
the `Prefix` type if a single slash is used as the pattern - ingress
expects that `/bar` would match, but it doesn't, matching only `/bar/`
instead. This works well with declared hostnames because HAProxy Ingress
builds the pattern concatenating hostname and path, but this doesn't
happen with the default hostname where the path is configured alone.

This update adds a prefix in the default hostname patterns, just like
it's already happening with declared hostnames, in order to them to
behave in the same way.