fix reload failure detection on 2.5+
HAProxy 2.5 and newer has a new lay out for the `show proc` command. The
compatibility change was recently implemented in the controller. Besides
the lay out, the way to identity if an instance was properly started was
changed as well. This update fixes the reading of this information. This
impacts only haproxy running as an external process, and only with
haproxy 2.5 and newer.

This update should be merged as far as v0.12, where external haproxy was
first implemented.