add suffix to name of local frontend proxies
Some frontends cannot belong to frontend metrics depending on controller
configurations. Actually we have `_front__tls` and `_front__auth` in
this scenario.

`_front__tls` is a TCP frontend created in front of the formal https
frontend. This new frontend is created whenever a `ssl-passthrough`
configuration is added to an ingress resource. It will send the request
straight to a backend or to the https front depending on the client's
SNI. In this scenario the `_front_tcp` metrics should be used and
`_front_https` should be discarded, otherwise metrics from
ssl-passthrough ingress wouldn't be added. `_front_https` should be
discarded because it's already being considered from `_front__tls`. On
the other hand `_front_https` should be added to the metric count if
`_front_tls` doesn't exist.

`_front__auth` is a HTTP frontend used by external authentication. It is
used by the Lua script that calls the backend that should authorize the
request. Although it is a frontend, its metrics are already being
added to the backends that connect to the authorization servers. We
could indeed create a new dashboard for it, but it'd be incorrect to
add its datas to the formal frontend metrics.

This update changes how the naming of these frontends are made, which
simplifies how the metrics dashboard is configured. `_front_https` will
receive a `__local` suffix if it is behind `_front_tls`. `_front__auth`
will always receive this suffix. The dashboard is configured to just
ignore `^.*__local$` when building frontend metrics.

This update might break backward compatibility, already described in
the release notes.