don't allow empty origins in cors-allow-origin
If the first origin (used for setting the header when origin doesn't match)
is an empty string, it causes haproxy config parsing to fail,
because `http-request set-header` requires two non-empty strings as arguments.

Also the Origin http header shouldn't contain empty string so there's
no reason to match for it.