check type assertion on all informers
Delete event from informers might return `cache.DeletedFinalStateUnknown`
instead of the actual object if the object is removed from the api while
the watch is closed, e.g. during an outage. Informers that need to check
the state of a removed object must check the type, otherwise a panic
would be issued. This update implements a safe typecast on the delete
event of all informers that were missing it.