add lock before call acmeCheck()
AcmeCheck() is an instance's method that can be triggered from three
distinct goroutines. Although it doesn't change the model, it reads it
and can lead to inconsistent data or even a race in the map of acme
related data if called concurrently with a model update. This commit
prevents that AcmeCheck() is called during instance's model updates.