remove readiness endpoint from starting config
External HAProxy starts with a bare minimum configuration, just to serve
as a placeholder to the first real configuration. We need this first
configuration so we have a process to be reloaded remotely. This needs
to be a bare minimum / invalid because in that stage we don't have
information enough to create a real configuration.

This starting configuration however might lead to a false positive
ready state from the cluster perspective because of the /healthz
endpoint implemented in the starting configuration. This endpoint is
being removed from that stage, so the cluster can property mark the
pod as not ready and remove from load balancers until the real
configuration is in place.