Documents the expected format for --configmap key
If this format is not followed following error is thrown:

```
2022/08/01 08:50:43 unable to parse static config: error reading global ConfigMap 'haproxy-ingress': an empty namespace may not be set when a resource name is provided
```
that actually is thrown by k8.

I'm not sure if it is a common knowledge to always use namespace when specifying a configmap. Ignore this if that is the case.
If not, then i guess the format can also be enforced at the time of parsing `--configmap` arg?